pyproject.toml
[tool.poetry]
name = "lessons-project"
version = "0.1.0"
description = ""
authors = ["spirin-vadim <vadis99@ya.ru>"]
readme = "README.md"
packages = [{include = "lessons_project"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
djangorestframework = "^3.14.0"
dj-rest-auth = "^5.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
